pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
]
